Manually set INSTALLDIR
